+++ ModelSwapperSkins/BoneMapping/InitializerRules/BoneInitializerRules_Croco.cs
@@ -1,4 +1,5 @@
 ﻿using RoR2;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace ModelSwapperSkins.BoneMapping.InitializerRules
@@ -27,6 +28,7 @@ namespace ModelSwapperSkins.BoneMapping.InitializerRules
                     boneInfo.RotationOffset *= Quaternion.Euler(0f, 180f, 0f);
                     break;
                 case BoneType.Head:
+                    boneInfo.PositionOffset += new Vector3(0f, 1f, -2f);
                     boneInfo.RotationOffset *= Quaternion.Euler(270f, 180f, 0f);
                     break;
                 case BoneType.Jaw:
@@ -46,6 +48,7 @@ namespace ModelSwapperSkins.BoneMapping.InitializerRules
                     break;
                 case BoneType.FootL:
                 case BoneType.FootR:
+                    boneInfo.PositionOffset += new Vector3(0f, 0.25f, 0f);
                     boneInfo.RotationOffset *= Quaternion.Euler(290f, 0f, 0f);
                     break;
                 case BoneType.ShoulderR:
@@ -67,6 +70,77 @@ namespace ModelSwapperSkins.BoneMapping.InitializerRules
             }
 
             return boneInfo;
+        }
+
+        public override IEnumerable<Bone> GetAdditionalBones(Transform modelTransform, List<Bone> existingBones)
+        {
+            foreach (Bone bone in base.GetAdditionalBones(modelTransform, existingBones))
+            {
+                yield return bone;
+            }
+
+            Bone neck1Bone = existingBones.Find(b => b.Info.Type == BoneType.Neck1);
+            if (neck1Bone != null)
+            {
+                yield return new Bone(neck1Bone)
+                {
+                    Info = new BoneInfo(BoneType.Chest)
+                    {
+                        PositionOffset = new Vector3(0f, -1.4142f, 1.4142f),
+                        RotationOffset = Quaternion.Euler(45f, 180f, 0f),
+                        MatchFlags = BoneMatchFlags.MatchToOther,
+                        ExclusionRules = [
+                            new BoneExclusionRule([BoneType.Neck1], OtherBoneMatchExclusionRuleType.ExcludeIfAnyMatch)
+                        ]
+                    }
+                };
+            }
+            else
+            {
+                Log.Error("Failed to find Neck1 bone");
+            }
+
+            Bone shoulderLBone = existingBones.Find(b => b.Info.Type == BoneType.ShoulderL);
+            if (shoulderLBone != null)
+            {
+                yield return new Bone(shoulderLBone)
+                {
+                    Info = new BoneInfo(BoneType.Chest)
+                    {
+                        PositionOffset = new Vector3(0f, -1.4142f, -1.4142f),
+                        RotationOffset = Quaternion.Euler(0f, 270f, 315f),
+                        MatchFlags = BoneMatchFlags.MatchToOther,
+                        ExclusionRules = [
+                            new BoneExclusionRule([BoneType.ShoulderL], OtherBoneMatchExclusionRuleType.ExcludeIfAnyMatch)
+                        ]
+                    }
+                };
+            }
+            else
+            {
+                Log.Error("Failed to find ShoulderL bone");
+            }
+
+            Bone shoulderRBone = existingBones.Find(b => b.Info.Type == BoneType.ShoulderR);
+            if (shoulderRBone != null)
+            {
+                yield return new Bone(shoulderRBone)
+                {
+                    Info = new BoneInfo(BoneType.Chest)
+                    {
+                        PositionOffset = new Vector3(0f, -1.4142f, -1.4142f),
+                        RotationOffset = Quaternion.Euler(0f, 90f, 45f),
+                        MatchFlags = BoneMatchFlags.MatchToOther,
+                        ExclusionRules = [
+                            new BoneExclusionRule([BoneType.ShoulderR], OtherBoneMatchExclusionRuleType.ExcludeIfAnyMatch)
+                        ]
+                    }
+                };
+            }
+            else
+            {
+                Log.Error("Failed to find ShoulderR bone");
+            }
         }
     }
 }